pytest: Migrate close and onchain tests to test_closing.py
